teasing out a better 'business logic' boundary
Now there's a single repository interface that handles all the side effects.
I've tried to design it to allow sensible extension points at the following
points in the computation:

* at the beginning, when deciding which build constitutes the 'cycle end'
* at the end, when deciding which build constitutes the 'cycle start'
* in between builds, when deciding which builds should logically be considered
  'linked' (in this case, only allowing successful builds that finished earlier
  than the current build)
* (eventually, though the conceptual model isn't in place yet) deciding how to
  filter or at least sort/prioritize amongst a set of paths from the same
  cycle start candidate and the cycle end - even in the case of the Concourse
  pipeline there are a number of paths from the `publish-binaries` build that
  shipped v5.8.0 back to various builds of `unit`, so maybe it makes sense to
  establish a selection strategy that operates not only on the 'cycle start
  candidate itself', but on the whole path or 'cycle candidate'.

Next steps will be to establish a reasonably-smart fake in-memory implementation
of the repository interface, which probably means eliminating our go-concourse
fakes and setting up an integration suite for the existing API repository.

I'd also like to extract interfaces for the above 'extension/decision points'
in the execution of the program, potentially extracting notions of 'filter'
which determine which candidates are unilaterally unacceptable, and 'priority'
which just orders a set of candidates for suitability when you can't get a
unique match.

Having done those two things I think it should be possible for the `LeadTime`
function itself to have strong unit tests with resilient inversion of
dependencies, as it's really the most core domain concept.

Following this I'd like to add a DB repository implementation, because i think
the limitations of what the API exposes may mean that it is not capable of
discovering as many links as an implementation with direct access to the raw
database structures -- plus direct DB access is bound to be a bit faster.

Signed-off-by: Jamie Klassen <cklassen@pivotal.io>